Remove tox config
It's silly to repeat the dependencies (not sure why they are not picked
up when tox installs the cachedjdk sdist, but they aren't). Also, I
can't find a way to install cachedjdk[testing] in tox. If necessary, we
can always generate a tox.ini on the fly from data in pyproject.toml.